fix fix branch

+++ src/main/java/Calculator.java
@@ -1,6 +1,6 @@
 public class Calculator {
 
-    public static double multiply(double a, double b ){
+    public static int multiply(int a, int b ){
         return a*b;
     }
 }
